requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-types==0.7.0
appdirs==1.4.4
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
beartype==0.15.0
bleach==6.2.0
bokeh==3.8.0
boto3==1.35.99
botocore==1.35.99
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
contourpy==1.3.3
decorator==5.2.1
deltalake==0.17.4
diskcache==5.6.3
executing==2.2.1
frozenlist==1.7.0
fs==2.4.16
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.3
google-cloud-pubsub==2.31.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
h3==4.3.1
httplib2==0.31.0
idna==3.10
importlib_metadata==8.7.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jupyter_bokeh==4.0.5
jupyterlab_widgets==3.0.15
linkify-it-py==2.0.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
multidict==6.6.4
narwhals==2.5.0
networkx==3.5
numpy==2.3.3
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pandas==2.3.2
panel==1.8.1
param==2.2.1
parso==0.8.5
pathway==0.26.1
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-sat==1.8.dev23
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.3
requests==2.32.5
rich==14.1.0
rsa==4.9.1
s3transfer==0.10.4
scikit-learn==1.7.2
scipy==1.16.2
shapely==2.1.2
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
widgetsnbextension==4.0.14
xyzservices==2025.4.0
yarl==1.20.1
zipp==3.23.0
google-genai==1.39.1
faiss-cpu==1.12.0
sentence-transformers==5.1.1
python-dotenv==1.1.1
